+++ FirebaseAdmin/FirebaseAdmin/Auth/FirebaseAuth.cs
@@ -264,6 +264,41 @@ namespace FirebaseAdmin.Auth
         }
 
         /// <summary>
+        /// Creates a new user account with the attributes contained in the specified <see cref="UserRecordArgs"/>.
+        /// </summary>
+        /// <param name="args">Attributes that will be added to the new user account.</param>
+        /// <returns>A task that completes with a <see cref="UserRecord"/> representing
+        /// the newly created user account.</returns>
+        /// <exception cref="ArgumentNullException">If <paramref name="args"/> is null.</exception>
+        /// <exception cref="ArgumentException">If any of the values in <paramref name="args"/> are invalid.</exception>
+        /// <exception cref="FirebaseException">If an error occurs while creating rhe user account.</exception>
+        public async Task<UserRecord> CreateUserAsync(UserRecordArgs args)
+        {
+            return await this.CreateUserAsync(args, default(CancellationToken))
+                .ConfigureAwait(false);
+        }
+
+        /// <summary>
+        /// Creates a new user account with the attributes contained in the specified <see cref="UserRecordArgs"/>.
+        /// </summary>
+        /// <param name="args">Attributes that will be added to the new user account.</param>
+        /// <param name="cancellationToken">A cancellation token to monitor the asynchronous
+        /// operation.</param>
+        /// <returns>A task that completes with a <see cref="UserRecord"/> representing
+        /// the newly created user account.</returns>
+        /// <exception cref="ArgumentNullException">If <paramref name="args"/> is null.</exception>
+        /// <exception cref="ArgumentException">If any of the values in <paramref name="args"/> are invalid.</exception>
+        /// <exception cref="FirebaseException">If an error occurs while creating rhe user account.</exception>
+        public async Task<UserRecord> CreateUserAsync(
+            UserRecordArgs args, CancellationToken cancellationToken)
+        {
+            var userManager = this.IfNotDeleted(() => this.userManager.Value);
+            var uid = await userManager.CreateUserAsync(args, cancellationToken)
+                .ConfigureAwait(false);
+            return await userManager.GetUserByIdAsync(uid, cancellationToken).ConfigureAwait(false);
+        }
+
+        /// <summary>
         /// Gets a <see cref="UserRecord"/> object containing information about the user who's
         /// user ID was specified in <paramref name="uid"/>.
         /// </summary>
@@ -274,7 +309,8 @@ namespace FirebaseAdmin.Auth
         /// <exception cref="FirebaseException">If a user cannot be found with the specified user ID.</exception>
         public async Task<UserRecord> GetUserAsync(string uid)
         {
-            return await this.GetUserAsync(uid, default(CancellationToken));
+            return await this.GetUserAsync(uid, default(CancellationToken))
+                .ConfigureAwait(false);
         }
 
         /// <summary>
@@ -293,7 +329,8 @@ namespace FirebaseAdmin.Auth
         {
             var userManager = this.IfNotDeleted(() => this.userManager.Value);
 
-            return await userManager.GetUserByIdAsync(uid, cancellationToken);
+            return await userManager.GetUserByIdAsync(uid, cancellationToken)
+                .ConfigureAwait(false);
         }
 
         /// <summary>
@@ -307,7 +344,8 @@ namespace FirebaseAdmin.Auth
         /// <exception cref="FirebaseException">If a user cannot be found with the specified email.</exception>
         public async Task<UserRecord> GetUserByEmailAsync(string email)
         {
-            return await this.GetUserByEmailAsync(email, default(CancellationToken));
+            return await this.GetUserByEmailAsync(email, default(CancellationToken))
+                .ConfigureAwait(false);
         }
 
         /// <summary>
@@ -326,7 +364,8 @@ namespace FirebaseAdmin.Auth
         {
             var userManager = this.IfNotDeleted(() => this.userManager.Value);
 
-            return await userManager.GetUserByEmailAsync(email, cancellationToken);
+            return await userManager.GetUserByEmailAsync(email, cancellationToken)
+                .ConfigureAwait(false);
         }
 
         /// <summary>
@@ -340,7 +379,8 @@ namespace FirebaseAdmin.Auth
         /// <exception cref="FirebaseException">If a user cannot be found with the specified phone number.</exception>
         public async Task<UserRecord> GetUserByPhoneNumberAsync(string phoneNumber)
         {
-            return await this.GetUserByPhoneNumberAsync(phoneNumber, default(CancellationToken));
+            return await this.GetUserByPhoneNumberAsync(phoneNumber, default(CancellationToken))
+                .ConfigureAwait(false);
         }
 
         /// <summary>
@@ -359,7 +399,8 @@ namespace FirebaseAdmin.Auth
         {
             var userManager = this.IfNotDeleted(() => this.userManager.Value);
 
-            return await userManager.GetUserByPhoneNumberAsync(phoneNumber, cancellationToken);
+            return await userManager.GetUserByPhoneNumberAsync(phoneNumber, cancellationToken)
+                .ConfigureAwait(false);
         }
 
         /// <summary>
@@ -371,7 +412,8 @@ namespace FirebaseAdmin.Auth
         /// <exception cref="FirebaseException">If an error occurs while deleting the user.</exception>
         public async Task DeleteUserAsync(string uid)
         {
-            await this.DeleteUserAsync(uid, default(CancellationToken));
+            await this.DeleteUserAsync(uid, default(CancellationToken))
+                .ConfigureAwait(false);
         }
 
         /// <summary>
@@ -387,7 +429,8 @@ namespace FirebaseAdmin.Auth
         {
             var userManager = this.IfNotDeleted(() => this.userManager.Value);
 
-            await userManager.DeleteUserAsync(uid, cancellationToken);
+            await userManager.DeleteUserAsync(uid, cancellationToken)
+                .ConfigureAwait(false);
         }
 
         /// <summary>
@@ -408,7 +451,8 @@ namespace FirebaseAdmin.Auth
         public async Task SetCustomUserClaimsAsync(
             string uid, IReadOnlyDictionary<string, object> claims)
         {
-            await this.SetCustomUserClaimsAsync(uid, claims, default(CancellationToken));
+            await this.SetCustomUserClaimsAsync(uid, claims, default(CancellationToken))
+                .ConfigureAwait(false);
         }
 
         /// <summary>
@@ -432,7 +476,7 @@ namespace FirebaseAdmin.Auth
             string uid, IReadOnlyDictionary<string, object> claims, CancellationToken cancellationToken)
         {
             var userManager = this.IfNotDeleted(() => this.userManager.Value);
-            var user = new UserArgs()
+            var user = new UserRecordArgs()
             {
                 Uid = uid,
                 CustomClaims = claims,